+++ vm/src/types/program.rs
@@ -32,8 +32,6 @@ use core::num::NonZeroUsize;
 use std::path::Path;
 
 use super::builtin_name::BuiltinName;
-#[cfg(feature = "extensive_hints")]
-use super::relocatable::Relocatable;
 #[cfg(feature = "test_utils")]
 use arbitrary::{Arbitrary, Unstructured};
 
@@ -109,10 +107,7 @@ impl<'a> Arbitrary<'a> for SharedProgramData {
 pub struct HintsCollection {
     pub hints: Vec<HintParams>,
     /// This maps a PC to the range of hints in `hints` that correspond to it.
-    #[cfg(not(feature = "extensive_hints"))]
     pub(crate) hints_ranges: Vec<HintRange>,
-    #[cfg(feature = "extensive_hints")]
-    pub hints_ranges: HashMap<Relocatable, HintRange>,
 }
 
 impl HintsCollection {
@@ -137,21 +132,15 @@ impl HintsCollection {
         }
 
         let mut hints_values = Vec::with_capacity(full_len);
-        #[cfg(not(feature = "extensive_hints"))]
         let mut hints_ranges = vec![None; max_hint_pc + 1];
-        #[cfg(feature = "extensive_hints")]
-        let mut hints_ranges = HashMap::default();
         for (pc, hs) in hints.iter().filter(|(_, hs)| !hs.is_empty()) {
             let range = (
                 hints_values.len(),
                 NonZeroUsize::new(hs.len()).expect("empty vecs already filtered"),
             );
-            #[cfg(not(feature = "extensive_hints"))]
             {
                 hints_ranges[*pc] = Some(range);
             }
-            #[cfg(feature = "extensive_hints")]
-            hints_ranges.insert(Relocatable::from((0_isize, *pc)), range);
             hints_values.extend_from_slice(&hs[..]);
         }
 
@@ -165,7 +154,6 @@ impl HintsCollection {
         self.hints.iter()
     }
 
-    #[cfg(not(feature = "extensive_hints"))]
     pub fn get_hint_range_for_pc(&self, pc: usize) -> Option<HintRange> {
         self.hints_ranges.get(pc).cloned()
     }
@@ -174,27 +162,19 @@ impl HintsCollection {
 impl From<&HintsCollection> for BTreeMap<usize, Vec<HintParams>> {
     fn from(hc: &HintsCollection) -> Self {
         let mut hint_map = BTreeMap::new();
-        #[cfg(not(feature = "extensive_hints"))]
         for (i, r) in hc.hints_ranges.iter().enumerate() {
             let Some(r) = r else {
                 continue;
             };
             hint_map.insert(i, hc.hints[r.0..r.0 + r.1.get()].to_owned());
         }
-        #[cfg(feature = "extensive_hints")]
-        for (pc, r) in hc.hints_ranges.iter() {
-            hint_map.insert(pc.offset, hc.hints[r.0..r.0 + r.1.get()].to_owned());
-        }
         hint_map
     }
 }
 
 /// Represents a range of hints corresponding to a PC as a  tuple `(start, length)`.
-#[cfg(not(feature = "extensive_hints"))]
 /// Is [`None`] if the range is empty, and it is [`Some`] tuple `(start, length)` otherwise.
-type HintRange = Option<(usize, NonZeroUsize)>;
-#[cfg(feature = "extensive_hints")]
-pub type HintRange = (usize, NonZeroUsize);
+pub type HintRange = Option<(usize, NonZeroUsize)>;
 
 #[cfg_attr(feature = "test_utils", derive(Arbitrary))]
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -469,7 +449,6 @@ impl TryFrom<CasmContractClass> for Program {
 #[cfg(test)]
 impl HintsCollection {
     pub fn iter(&self) -> impl Iterator<Item = (usize, &[HintParams])> {
-        #[cfg(not(feature = "extensive_hints"))]
         let iter = self
             .hints_ranges
             .iter()
@@ -484,15 +463,6 @@ impl HintsCollection {
                     }
                 })
             });
-        #[cfg(feature = "extensive_hints")]
-        let iter = self.hints_ranges.iter().filter_map(|(pc, (start, len))| {
-            let end = start + len.get();
-            if end <= self.hints.len() {
-                Some((pc.offset, &self.hints[*start..end]))
-            } else {
-                None
-            }
-        });
         iter
     }
 }
@@ -648,7 +618,6 @@ mod tests {
         assert_eq!(program.shared_program_data.main, None);
         assert_eq!(program.shared_program_data.identifiers, HashMap::new());
 
-        #[cfg(not(feature = "extensive_hints"))]
         let program_hints: HashMap<_, _> = program
             .shared_program_data
             .hints_collection
@@ -660,19 +629,6 @@ mod tests {
                 (
                     pc,
                     program.shared_program_data.hints_collection.hints[s..e].to_vec(),
-                )
-            })
-            .collect();
-        #[cfg(feature = "extensive_hints")]
-        let program_hints: HashMap<_, _> = program
-            .shared_program_data
-            .hints_collection
-            .hints_ranges
-            .iter()
-            .map(|(pc, (s, l))| {
-                (
-                    pc.offset,
-                    program.shared_program_data.hints_collection.hints[*s..(s + l.get())].to_vec(),
                 )
             })
             .collect();